+++ shapes/Shape.h
@@ -18,7 +18,7 @@ public:
 
     void setParam1(int param1);
     void setParam2(int param2);
-    void setParam3(int param3);
+    void setParam3(float param3);
 
     void updateGL(GLuint shader);
     void updateNormals(NormalRenderer *normRenderer);
@@ -34,21 +34,21 @@ protected:
     void makeRect(int *i, struct Rect *rect);
 
     void printVert(glm::vec3 v);
-    void setParamMax(int p1, int p2, int p3);
+    void setParamMax(int p1, int p2, float p3);
 
     GLuint m_vaoID; // The vertex array object ID, which is passed to glBindVertexArray.
     GLfloat *m_vertexData;
 
     int m_p1;
     int m_p2;
-    int m_p3;
+    float m_p3;
 
     int m_numVerts;
 
 private:
     int m_p1max;
     int m_p2max;
-    int m_p3max;
+    float m_p3max;
 };
 
 #endif // SHAPE_H
